Fix syntax error in _config.yml
YML format is basically a set of key-value pairs where key and value a
separated by `:`. Therefore if the key or the value contain `:` we need
to escape them to not confuse the parser that might interpret `:` in the
key/value as a separator.

This change escapes values in the _config.yml to escape `:` used in the
page description.