Add a post introducing the FTDI USB cable and skeleton of the USB driver
The post covers the basics of FTDI library and how to use it to override
the default VID and PID of the device.

Additionally the post covers VID and PID concepts and how they are used
in Linux systems to detect the right driver for the device and even
automatically load it when the device is plugged in.

Finally the post shows the skeleton of the USB driver that handles
devices with a particual VID and PID. The driver does not do anything
useful at this point though.