Add a post covering USB data transfers and I2C adapters
This is the final post for my hobby project on creating an USB-to-I2C
adapter using an FTDI-based cabel with MPSSE inside. The post covers
simple syncrhonous control and bulk data transfers as well as
registering an I2C adapter with the kernel.

I did not provide the complete code in the post and only covered some
small code snippets to explain how to organize USB data transfers and
register an I2C controller with the kernel.

The complete code is available on https://github.com/krinkinmu/bootlin,
so there is no need to repeat it multiple times.