[Work-In-Progress] Add an example of math equations on github pages
What seem to be required here is to add HTML script clause to the layout
to download MathJax scripts. That seem to be all that is needed, but I'm
yet to verify that.

I created a separate layout for posts with math that inherits the
default layout. Given that the layout is an HTML template, it's not
entirely clear what the inheritance will produce in the end though.