Cover matrix form of the recurrence relations in the post
This change adds a section about writing the Fibonacci recurrence
relation in a matrix form. A long the way as an interesting side note
this change references a cool (though from practial point of view not
very useful) method of calculating Fibonacci numbers by efficiently
calculating the N-th power of the matrix.