Add a post about running EFI app on ARM
The post covers how to change the builing options to build the EFI
binary and ELF binary for aarch64. There was a change to the actual code
required to make it build for aarch64, but it was a rather minor
modification.

The post also covers how to test the binary on HiKey960 (an actually
available phisical aarch64 board with firmware supporint UEFI
specification) and on QEMU.