Add a post covering PL011 communication
It covers some basics of UART, PL011 specification and how to program
the device, how to find the configuration parameters from the DT and
test the change in QEMU.

While writing this post I discovered that the most recent change to the
EFI repositry actually breaks loading ELF binaries at least on AARCH64,
so I reverted the change for now. Something is deeply magical going on
there, so I need more time to debug the issue than I have right now.