Add a post about my attempt to try to make PL011 work on my HiKey960
This post covers the parameters of the UART for HiKey960 board I have.
It also gives a brief recollection of me debugging why the code I had
for PL011 didn't work with the actual hardware board.

I cannot claim that I solved the problem, after all I would expect that
reset should work in any hardware device state, but I also could
actually pin point what is that exactly that I'm doing wrong in the code
that violates the specification.