Add an introductionary post covering AArch64 EL basics
The post contains some stuff that I wanted to cover while explaining
interrupt handling in AArch64 but was deemed to make the post on
interrupt handling too long.

So what this post contains is:

  * general idea of AArch64 privildge levels
  * registers involved into interrupt handling SPSR and ELR
  * jumping between ELs
  * basic ARM assembly

So it's quite a mix of stuff, so probably it's not structured in the
best way possible.