Add a post on interrupt and exception handling in AArch64
This post builds on top of the previous post and gives and example of
configuring interrupt routing, setting up exception vector table and
gives and example of an interrupt handling routine.

The interrupt handling routine is however not particularly useful to
explain how the interrupts and exceptions should be handled in general,
but it should be useful to explain the framework where the actual
handling logic should fit in.