Add a post about an interesting probability problem
The post touches on the Bayes Theorem and The Law of Total Probability.
The problem setting is quite simple and involves coins (one of which
unfair).

I additionally included Python code that simulates the experiment to get
some additional confidence in the results of the calculations. Plus it
should give an idea of why simulations might be quite inneficient.