A short post on U-boot standard boot process
It covers the basics of the U-boot standard boot logic, how to create
a boot script and where to put it, the order of boot devices and boot
methods.

I included references to the U-boot source code as well for a good
measure and for readers interested in poking into the code.